refactor: extract sidebar helper types and consolidate Context handling
Continue the ConnectDatasetSidebar readability refactor:

- Extract ExperimentEntry and DatasetsConnectedEvent from their nested
  position in ConnectDatasetSidebar into first-class, same-package files
  so the sidebar's public API and data types are discoverable.
- Introduce a single currentProjectId() accessor that centralises the
  repeated 'context != null && projectId present' null/emptiness checks
  previously scattered across setContext, open, loadExperimentsAsync and
  connectSelectedDatasets.
- Drop the now-unused ExperimentId and NonNull imports.

Behavior-preserving: the extracted types are package-visible in the same
package and referenced only implicitly (via addDatasetsConnectedListener),
so no external call sites or the public event contract change.
